+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Feature - data cleaning history api
 ## Current Release 
 ### 0.209.0 
 **Release Date:** Fri Aug 31 02:33:32 UTC 2018